Better merging and assorted refactoring
