+++ client/src/components/includes/SelectedTags.tsx
@@ -5,7 +5,7 @@ class SelectedTags extends React.Component {
     props: {
         index: number,
         tag: string,
-        ifSelected: boolean,
+        isSelected: boolean,
         onClick: Function | null,
         level: number
     };
@@ -22,17 +22,16 @@ class SelectedTags extends React.Component {
     }
 
     render() {
-        var baseClassName = '';
-        if (this.props.level === 1) {
-            baseClassName = 'tag primaryTag';
-        } else if (this.props.level === 2) {
-            baseClassName = 'tag secondaryTag';
-        }
-        if (this.props.ifSelected) {
-            return <p className={baseClassName + ' selectedTag'} onClick={this._onClick}>{this.props.tag}</p>;
-        } else {
-            return <p className={baseClassName} onClick={this._onClick}>{this.props.tag}</p>;
-        }
+        return (
+            <p
+                className={'tag ' +
+                    (this.props.level === 1) ? 'primaryTag' : 'secondaryTag' +
+                    this.props.isSelected && ' selectedTag'}
+                onClick={this._onClick}
+            >
+                {this.props.tag}
+            </p>
+        );
     }
 }
