+++ client/src/components/includes/SessionInformationHeader.tsx
@@ -23,7 +23,7 @@ class SessionInformationHeader extends React.Component {
             return (
                 <header className="DesktopSessionInformationHeader" >
                     <div className="Picture">
-                        <img src={session.sessionTasBySessionId.nodes[0].userByUserId.photoUrl} />
+                        <img src={session.sessionTasBySessionId.nodes[0].userByUserId.photoUrl || avatar} />
                     </div>
                     <div className="Details">
                         <p className="Location">{session.building + ' ' + session.room}</p>
@@ -65,7 +65,7 @@ class SessionInformationHeader extends React.Component {
                             <Moment date={session.endTime} interval={0} format={' - h:mm A'} />
                         </div>
                         <div className="Picture">
-                            <img src={session.sessionTasBySessionId.nodes[0].userByUserId.photoUrl} />
+                            <img src={session.sessionTasBySessionId.nodes[0].userByUserId.photoUrl || avatar} />
                         </div>
                     </div>
                 </div>
