--- client/src/components/includes/SessionPopularQuestion.tsx
@@ -1,24 +1,0 @@
-import * as React from 'react';
-
-class SessionPopularQuestion extends React.Component {
-
-    props: {
-        question: string,
-        numPeople: number
-    };
-
-    render() {
-        return (
-            <div className="SessionPopularQuestion">
-                <div className="SessionPopularQuestion-Text">
-                    {this.props.question}
-                </div>
-                <div className="SessionPopularQuestion-People">
-                    n: {this.props.numPeople}
-                </div>
-            </div>
-        );
-    }
-}
-
-export default SessionPopularQuestion;
