+++ client/src/components/pages/LoginView.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 const QMeLogo = require('../../media/QMeLogo.svg');
 const QLogo = require('../../media/QLogo.svg');
-import LoginButton from '../includes/LoginButton';
+const googleLogo = require('../../media/googleLogo.svg');
 
 class LoginView extends React.Component {
     render() {
@@ -13,7 +13,10 @@ class LoginView extends React.Component {
                 </section>
                 <section className="bottomPanel">
                     <p className="hintText" >Use your Cornell NetID to login</p>
-                    <LoginButton URL="/__auth" />
+                    <a className="loginButton" href="/__auth" >
+                        <img src={googleLogo} className="googleLogo" />
+                        <span className="loginButtonText">Sign in with Google</span>
+                    </a>
                     <img src={QLogo} className="QLogo" />
                 </section>
             </div >
